+++ packages/zenrouter_devtools/example/android/app/src/main/kotlin/dev/zennn/router/example/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package dev.zennn.router.example.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()